requirements.txt
# See also setup.py install_requires
#
stem==1.8.0
aiohttp==3.6.2
aiohttp-jinja2==1.2.0
ckcc-protocol>=1.0.0
pynacl==1.3.0
aiohttp_session
click
pendulum
pyyaml
Pillow
pytest
requests[socks]
